style(scpi): pre-PR review cleanups in AdcChannelResolve
Three findings from the code-review pass, none behavioural:

* dropped the redundant `(size_t)` cast on the bound test -- AInArray::Size
  is already size_t (ArrayWrapper.h:20), and the comment now states the
  property the test actually relies on: both operands are unsigned, so the
  (size_t)-1 that ADC_FindChannelIndex returns for an id with no entry is the
  largest representable value and fails the test;
* reflowed the ragged lines the AdcChannelArgInRange header edit left behind;
* no other change.

The built hex is BYTE-IDENTICAL to the image the bench evidence was gathered
on (md5 bcf91feeb91fb7ed5ba26eb1f9437560, firmware_crc32 4B2F83C1), verified
by rebuilding and comparing rather than assumed -- so the acceptance run and
the mutation proof still describe exactly this source.

Also re-verified in this pass: all NINE AdcChannelResolve call sites are
preceded by a range guard (eight by AdcChannelArgInRange, the two-arg
CONFigure:ADC:CHANnel by its own inline `param1 < 0 || param1 > 255` at
SCPIADC.c:357), which is the precondition the helper's (uint8_t) cast needs;
and the diff adds and removes no `.pattern` entry, so the SCPI wiki-sync gate
has nothing to reconcile.

Refs #888

+++ firmware/src/services/SCPI/SCPIADC.c
@@ -40,9 +40,8 @@
 // which handles the sparse id space -- AdcChannelResolve below, which #888
 // made the one owner of that second test (it was open-coded, three different
 // ways, at all nine sites). That space is sparse on BOTH variants: user ids
-// 0..15 on NQ1 and 0..7 on
-// NQ3, plus the same monitoring block 248..255 on each (ADC_CHANNEL_3_3V ..
-// ADC_CHANNEL_5VREF, AInConfig.h:240-247, pulled in by
+// 0..15 on NQ1 and 0..7 on NQ3, plus the monitoring block 248..255 on each
+// (ADC_CHANNEL_3_3V .. ADC_CHANNEL_5VREF, AInConfig.h:240-247, pulled in by
 // COMMON_MONITORING_CHANNELS_BOARDCONFIG). Widening this test to the
 // per-variant user maximum would usurp that handling, which is the regression
 // #682 had to undo.
@@ -144,7 +143,10 @@ static bool AdcChannelResolve(scpi_t * context, int channel, const char * cmd,
             0);
     size_t resolved = ADC_FindChannelIndex((uint8_t) channel);
 
-    if (resolved < (size_t) pBoardConfigAInChannels->Size) {
+    /* Size is already size_t (ArrayWrapper.h:20), so no cast: both operands
+     * are unsigned and (size_t)-1, what ADC_FindChannelIndex returns for an id
+     * with no entry, is the largest representable value and fails this test. */
+    if (resolved < pBoardConfigAInChannels->Size) {
         *index = resolved;
         return true;
     }
